chore: move pathmap to cache
