+++ ethers-solc/src/cache.rs
@@ -1,13 +1,13 @@
 //! Support for compiling contracts
 use crate::{
-    artifacts::Sources,
+    artifacts::{Contracts, Sources},
     config::SolcConfig,
     error::{Result, SolcError},
     utils, ArtifactOutput,
 };
 use serde::{Deserialize, Serialize};
 use std::{
-    collections::BTreeMap,
+    collections::{BTreeMap, HashMap},
     fs::{self, File},
     path::{Path, PathBuf},
     time::{Duration, UNIX_EPOCH},
@@ -284,6 +284,64 @@ impl CacheEntry {
     /// Returns the time
     pub fn last_modified(&self) -> Duration {
         Duration::from_millis(self.last_modification_date)
+    }
+}
+
+/// A helper type to handle source name/full disk mappings
+///
+/// The disk path is the actual path where a file can be found on disk.
+/// A source name is the internal identifier and is the remaining part of the disk path starting
+/// with the configured source directory, (`contracts/contract.sol`)
+#[derive(Debug, Default)]
+pub struct PathMap {
+    /// all libraries to the source set while keeping track of their actual disk path
+    /// (`contracts/contract.sol` -> `/Users/.../contracts.sol`)
+    pub source_name_to_path: HashMap<PathBuf, PathBuf>,
+    /// inverse of `source_name_to_path` : (`/Users/.../contracts.sol` -> `contracts/contract.sol`)
+    pub path_to_source_name: HashMap<PathBuf, PathBuf>,
+    /* /// All paths, source names and actual file paths
+     * paths: Vec<PathBuf> */
+}
+
+impl PathMap {
+    fn apply_mappings(sources: Sources, mappings: &HashMap<PathBuf, PathBuf>) -> Sources {
+        sources
+            .into_iter()
+            .map(|(import, source)| {
+                if let Some(path) = mappings.get(&import).cloned() {
+                    (path, source)
+                } else {
+                    (import, source)
+                }
+            })
+            .collect()
+    }
+
+    /// Returns all contract names of the files mapped with the disk path
+    pub fn get_artifacts(&self, contracts: &Contracts) -> Vec<(PathBuf, Vec<String>)> {
+        contracts
+            .iter()
+            .map(|(path, contracts)| {
+                let path = PathBuf::from(path);
+                let file = self.source_name_to_path.get(&path).cloned().unwrap_or(path);
+                (file, contracts.keys().cloned().collect::<Vec<_>>())
+            })
+            .collect()
+    }
+
+    pub fn extend(&mut self, other: PathMap) {
+        self.source_name_to_path.extend(other.source_name_to_path);
+        self.path_to_source_name.extend(other.path_to_source_name);
+    }
+
+    /// Returns a new map with the source names as keys
+    pub fn set_source_names(&self, sources: Sources) -> Sources {
+        Self::apply_mappings(sources, &self.path_to_source_name)
+    }
+
+    /// Returns a new map with the disk paths as keys
+    pub fn set_disk_paths(&self, sources: Sources) -> Sources {
+        Self::apply_mappings(sources, &self.source_name_to_path)
     }
 }
